Break into modules, add distribute command
- adds commands dir
- adds library module
- fixes some global vars
- adds distribute command
